perf(rsc): isolate page cache key marker
