+++ packages/vinext/src/server/app-page-probe.ts
@@ -1,5 +1,5 @@
 import { Fragment, isValidElement, type ReactElement, type ReactNode } from "react";
-import { markAppPagePropsForUseCache } from "vinext/shims/cache-runtime";
+import { markAppPagePropsForUseCache } from "vinext/shims/internal/app-page-props-cache-key";
 import { isNextRouterError } from "vinext/shims/navigation";
 import { collectAppPageSearchParams } from "./app-page-head.js";
 import {
